Add sleepless benchmark
This benchmark tests the cpu usage of waiting for threading.Condition
with a timeout. Built-in Condition in Python 2 is implemented using
polling and sleeps, which can use significant cpu time when lots of
threads are waiting.

cthreading uses the same design as in Python 3, waiting on a blocked
semaphore using sem_timedwait, which cost no cpu time.

pthreading uses pthread_cond_timedwait(). This introduces undefined
behavior if the condition is not used properly, and the issue of
spurious wakeups, which do not exists in sem_timedwait().