Raise thread.error when releasing un-acquired lock
Multiprocessing regression tests fail because cthreading was raising
RuntimeError when un-acquired lock is released (like Python 3).

Now we import the thread module and store thread.error in the module, so
we can raise it in this case.

Because we must import thread module, the check for thread module in
cthreding.monkeypatch() had to be removed.