Initialize waitq when condition is initialized
When condition is created, Python generic alloc function initialize the
memory to 0, but when condition is initialized after fork, waitq may
contain stale waiters for threads which do not exists in the child
process. This may cause notify to fail, trying to wake up a non-existing
thread, and keeping the main thread waiting forever on the condition.